move tests from unclear named test file
